Code now in a compilable, executable state for Win32 and Linux/Cygwin/MinGW.
I've been debugging the round execution code, and there is definitely something messed up with the pointers somewhere. The code can correctly find and execute an ability deemed to be the best (... 1>1?), but it doesn't seem to properly affect the correct character. I'm guessing it's hitting the incorrect Character pointers.

Signed-off-by: Johannes Lindhart Borresen <johannes@the.homestead.dk>