Possibly broken build.
The code compiles without issue but breaks down during the AbilityTable cycle for the Biggi character. Most likely this is due to a bad Character composition (targets on his Abilities), but I'm not sure.

In all cases, everything up to running should work, and running might even work with proper chars.

Signed-off-by: Johannes Lindhart Borresen <johannes@the.homestead.dk>